Apply bandwidth restrictions on both side of connection
When this.restrictBandwidth is set (on line 10206) to a non-zero value,
this is used in handleOffer to add a limit on the bandwidth the video
channel is allowed to use (only when useJingle is true).  There is no
corresponding limit in handleAnswer, however, so the channel proposer
does not have any bandwidth limit applied.

This change adds the same restriction to that side of the channel.

One can test this by changing

@@ -10203,7 +10203,7 @@ function PeerConnection(config, constraints) {
         });
     }
     // EXPERIMENTAL FLAG, might get removed without notice
-    this.restrictBandwidth = 0;
+    this.restrictBandwidth = 10;
     if (constraints && constraints.optional) {
         constraints.optional.forEach(function (constraint) {
             if (constraint.andyetRestrictBandwidth) {
@@ -10312,7 +10312,7 @@ function PeerConnection(config, constraints) {
         sid: '',
         isInitiator: true,
         sdpSessionID: Date.now(),
-        useJingle: false
+        useJingle: true
     };

     this.iceCredentials = {

and visually observing the reduced transmission quality on both sides
of a two-party chat (vs only affecting one side before this change).